# app/Swagger/Schemas/Common/Error401.php
<?php

namespace App\Swagger\Schemas\Common;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'Error401',
    properties: [
        new OA\Property(property: 'success', type: 'boolean', example: false),
        new OA\Property(property: 'message', type: 'string', example: 'Non authentifié. Veuillez vous connecter.'),
    ]
)]
class Error401 {}