# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion RH API — ARTF",
        "description": "API REST de gestion RH (intégration administrative des agents).\n\n## Authentification\n\n1. Appeler `POST /api/login` avec email / mot de passe\n2. Copier le `token` retourné\n3. Cliquer **Authorize** en haut à droite\n4. Saisir : `Bearer {token}` (ou uniquement le token selon le client)\n\nComptes de test (seeders) :\n- `admin@arft.cg` / `Admin@2026`\n- `rh@arft.cg` / `Rh@2026`\n\n## Convention de réponse\n\n```json\n{ \"data\": {}, \"message\": \"...\" }\n```\n\nErreurs : `{ \"success\": false, \"message\": \"...\", \"errors\": {} }`",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API (variable L5_SWAGGER_CONST_HOST)"
        }
    ],
    "paths": {
        "/api/integration/dossiers/{dossier}/actes": {
            "get": {
                "tags": [
                    "Intégration — Actes"
                ],
                "summary": "Lister les actes d'un dossier",
                "operationId": "listActesDossier",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ActeAdministratif"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intégration — Actes"
                ],
                "summary": "Générer un acte administratif (manuel)",
                "operationId": "genererActeAdministratif",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_acte"
                                ],
                                "properties": {
                                    "type_acte": {
                                        "type": "string",
                                        "example": "decision_recrutement"
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActeAdministratifResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/actes/{acte}/signer": {
            "post": {
                "tags": [
                    "Intégration — Actes"
                ],
                "summary": "Signer un acte administratif",
                "operationId": "signerActeAdministratif",
                "parameters": [
                    {
                        "name": "acte",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActeAdministratifResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error401": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié. Veuillez vous connecter."
                    }
                },
                "type": "object"
            },
...
            "Error422": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Le champ email est obligatoire."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "ActeAdministratif": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "dossier_integration_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_acte": {
                        "type": "string",
                        "example": "decision_recrutement"
                    },
                    "type_acte_label": {
                        "type": "string",
                        "example": "Décision de recrutement"
                    },
                    "numero": {
                        "type": "string",
                        "example": "ACTE-2026-0001"
                    },
                    "contenu": {
                        "type": "string",
                        "nullable": true
                    },
                    "fichier_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "signe": {
                        "type": "boolean",
                        "example": false
                    },
                    "date_signature": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ActeAdministratifResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/ActeAdministratif"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu via POST /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}